+++ tools/polyfill-ie11/index.ts
@@ -1,6 +1,7 @@
 import "core-js/modules/es6.promise";
-import "core-js/modules/es6.array.iterator.js";
-import "core-js/modules/es6.array.from.js";
-import "core-js/modules/es6.array.fill.js";
+import "core-js/modules/es6.array.iterator";
+import "core-js/modules/es6.array.from";
+import "core-js/modules/es6.array.fill";
+import "core-js/modules/es6.object.assign";
 import "es6-map/implement";
 import "whatwg-fetch";